clean up (part 5): moving classes to better location
