+++ plugins/org.eclipse.n4js.ide/META-INF/MANIFEST.MF
@@ -32,11 +32,11 @@ Export-Package: org.eclipse.n4js.ide,
  org.eclipse.n4js.ide.contentassist.antlr.internal,
  org.eclipse.n4js.ide.contentassist.antlr.lexer,
  org.eclipse.n4js.ide.editor.contentassist,
+ org.eclipse.n4js.ide.imports,
  org.eclipse.n4js.ide.server,
  org.eclipse.n4js.ide.server.codeActions.util,
  org.eclipse.n4js.ide.server.commands,
  org.eclipse.n4js.ide.server.hover,
- org.eclipse.n4js.ide.server.imports,
  org.eclipse.n4js.ide.xtext.server,
  org.eclipse.n4js.ide.xtext.server.build,
  org.eclipse.n4js.ide.xtext.server.concurrent
