+++ plugins/org.eclipse.n4js.ide/src/org/eclipse/n4js/ide/editor/contentassist/N4JSIdeContentProposalProvider.java
@@ -12,7 +12,6 @@ package org.eclipse.n4js.ide.editor.contentassist;
 
 import org.eclipse.emf.ecore.EObject;
 import org.eclipse.emf.ecore.EReference;
-import org.eclipse.n4js.ide.editor.contentassist.imports.ImportsAwareReferenceProposalCreator;
 import org.eclipse.n4js.n4JS.JSXElement;
 import org.eclipse.n4js.n4JS.ParameterizedPropertyAccessExpression;
 import org.eclipse.n4js.services.N4JSGrammarAccess;
