+++ plugins/org.eclipse.n4js.ide/src/org/eclipse/n4js/ide/server/codeActions/N4JSQuickfixProvider.java
@@ -21,9 +21,9 @@ import org.eclipse.lsp4j.Diagnostic;
 import org.eclipse.lsp4j.Position;
 import org.eclipse.lsp4j.Range;
 import org.eclipse.lsp4j.TextEdit;
-import org.eclipse.n4js.ide.editor.contentassist.imports.ImportsAwareReferenceProposalCreator;
-import org.eclipse.n4js.ide.editor.contentassist.imports.ReferenceResolution;
-import org.eclipse.n4js.ide.editor.contentassist.imports.ReferenceResolutionHelper;
+import org.eclipse.n4js.ide.editor.contentassist.ImportsAwareReferenceProposalCreator;
+import org.eclipse.n4js.ide.imports.ReferenceResolution;
+import org.eclipse.n4js.ide.imports.ReferenceResolutionHelper;
 import org.eclipse.n4js.ide.server.codeActions.util.ChangeProvider;
 import org.eclipse.n4js.n4JS.N4FieldDeclaration;
 import org.eclipse.n4js.n4JS.N4JSPackage;
