+++ plugins/org.eclipse.n4js.ide/src/org/eclipse/n4js/ide/server/commands/N4JSCommandService.java
@@ -42,13 +42,13 @@ import org.eclipse.lsp4j.TextEdit;
 import org.eclipse.lsp4j.WorkspaceEdit;
 import org.eclipse.n4js.external.LibraryChange;
 import org.eclipse.n4js.external.NpmCLI;
-import org.eclipse.n4js.ide.editor.contentassist.imports.ReferenceResolution;
-import org.eclipse.n4js.ide.editor.contentassist.imports.ReferenceResolutionHelper;
+import org.eclipse.n4js.ide.imports.ImportDescriptor;
+import org.eclipse.n4js.ide.imports.ImportOrganizer;
+import org.eclipse.n4js.ide.imports.ReferenceResolution;
+import org.eclipse.n4js.ide.imports.ReferenceResolutionHelper;
 import org.eclipse.n4js.ide.server.codeActions.ICodeActionAcceptor;
 import org.eclipse.n4js.ide.server.codeActions.N4JSCodeActionService;
 import org.eclipse.n4js.ide.server.codeActions.N4JSSourceActionProvider;
-import org.eclipse.n4js.ide.server.imports.ImportOrganizer;
-import org.eclipse.n4js.ide.server.imports.ImportDescriptor;
 import org.eclipse.n4js.ide.xtext.server.ExecuteCommandParamsDescriber;
 import org.eclipse.n4js.ide.xtext.server.XDocument;
 import org.eclipse.n4js.ide.xtext.server.XLanguageServerImpl;
